sign up

+++ server/views.py
@@ -5,9 +5,14 @@ from cryptography import x509
 from cryptography.x509.oid import NameOID
 import rsa
 import json
+from os import path
+from .certif_authority import CA
+from .db import Database
+import bcrypt
 
 
-
+ca = CA()
+db = Database()
 
 @api_view(['POST'])
 def register(request):
@@ -16,12 +21,15 @@ def register(request):
     body_unicode = request.body.decode('utf-8')
     user = json.loads(body_unicode)
     
-    #verify if user exists or not in LDAP ???
+    #verify if user exists or not
+    if(path.isfile(user["username"]+"_cert.pen") and path.exists(user["username"]+"_cert.pen")):
+        return Response(False)
+
 
     #generate private key
     public_key, private_key = rsa.newkeys(2048)
 
-    #save private key TO USER DEVICE ????????? OR ENCRYPT
+    #save private key TO USER
     with open(user["username"]+".pem", "wb") as f:
         f.write(private_key.save_pkcs1("PEM"))
 
@@ -34,15 +42,14 @@ def register(request):
     x509.NameAttribute(NameOID.COMMON_NAME, u"INSAT"),
     ]))
 
-    #send csr to certificate authority and get certificate ?????
-    
+    #send csr to certificate authority
+    ca.createCertificate(csr, user["username"])
 
-    #Save certif in client device??????
-    LDAPServer().add(user)
-    LDAPServer().connect()
+    password = user["password"]
+    hashed_password = bcrypt.hashpw(password.encode(), bcrypt.gensalt())
 
-    #Save in LDAP user infos + certif + pub key??????
-
+    #save user
+    db.execute("INSERT INTO users (name, email, username, password) VALUES (%s, %s, %s, %s)", (user["name"], user["email"],user["username"],hashed_password))
 
     return Response(True)
     
@@ -55,58 +62,18 @@ def login(request):
     user = json.loads(body_unicode)
 
     #verify if login data are valid
+    username = user["username"]
+    userFromDB = db.execute("SELECT * FROM users WHERE username like %s",(user["username"],))
 
-    #get certificate signature
-    with open(user["username"]+"_cert.pem", "rb") as cert_file:
-        cert_data = cert_file.read()
-    cert = x509.load_pem_x509_certificate(cert_data)
-    signature = cert.signature
+    print(userFromDB)
 
-    #verify signature
-    
-
-    return Response("login")
-
-
-from ldap3 import Server, Connection, ALL
-
-# Define server connection details
-ldap_server = "ldap://192.168.0.102:389"
-username = "cn=ines,dc=maxcrc,dc=com"
-password = "ines"
-
-class LDAPServer():
-
-    def connect(self):
-        server = Server(ldap_server)
-        conn = Connection(server, user=username, password=password)
-        conn.bind()
-        
-
-    # Search for an entry
-        base_dn = "dc=maxcrc,dc=com"
-        search_filter = "(cn=ines)"
-        result = conn.search(search_base=base_dn, search_filter=search_filter)
-    # Print the result
-        print(result)
-
-
-    def add(self,user):
-        server = Server(ldap_server)
-        
-        conn = Connection(server, user=username, password=password)
-        attributes = {
-            'objectClass': ['person', 'organizationalPerson', 'inetOrgPerson'],
-            'cn': user["username"],
-            'sn': user["lastName"],
-            'givenName': user["firstName"],
-            'mail': user["email"],
-        }
-        
-        conn.add("cn="+user["username"], attributes=attributes)
-        
-        if not conn.result['result'] == 0:
-            print(conn.result)
+    if(userFromDB!= None):
+        if (bcrypt.checkpw(user["password"].encode(), userFromDB[3].encode())):
+        #verify signature
+            resp = ca.verifyCertificate(user["username"])
+            return Response(resp)
         else:
-            print('Successfully added entry')
-
+            return Response(False)
+    else:
+        return Response(False)
+
